test: add external connections tests
