+++ src/index.test.ts
@@ -509,69 +509,3 @@ test('should be able validate fiatRates for all mainnet assets', async () => {
         expect(fiatRate).toBeGreaterThan(0);
     }
 })
-
-test('should be able validate externalConnections & forgetDappConnections', async () => {
-    const wallet = await setupWallet(defaultWalletOptions);
-    await wallet.dispatch.createWallet({
-        key: '0x1234567890123456789012345678901234567890',
-        mnemonic: 'test',
-        imported: true,
-    });
-    await wallet.dispatch.unlockWallet({
-        key: '0x1234567890123456789012345678901234567890',
-    });
-    expect(wallet.state.wallets.length).toBe(1);
-    expect(wallet.state.wallets[0].imported).toBe(true);
-    expect(wallet.state.unlockedAt).not.toBe(0);
-
-    await wallet.dispatch.initializeAnalyticsPreferences({
-        accepted: true,
-    });
-
-    expect(wallet.state.analytics.userId).not.toBe(null);
-    expect(wallet.state.analytics.acceptedDate).not.toBe(0);
-    expect(wallet.state.analytics.askedDate).not.toBe(0);
-    expect(wallet.state.analytics.askedTimes).toBe(0);
-    expect(wallet.state.analytics.notAskAgain).toBe(false);
-
-    const walletId = wallet.state.activeWalletId;
-    const mainnetAccounts = wallet?.state?.enabledAssets?.mainnet?.[walletId];
-    expect(mainnetAccounts).not.toBeNull();
-
-    // update balance this will generate address
-    if (mainnetAccounts) {
-        await wallet.dispatch.updateBalances({
-            network: Network.Mainnet,
-            walletId: walletId,
-            assets: mainnetAccounts,
-        });
-    }
-
-    const account = wallet.state.accounts?.[walletId]?.mainnet?.[1];
-    expect(account?.chain).toBe(ChainId.Ethereum);
-    const ethAccountId = account?.id;
-    const ethereumAddress = account?.addresses[0];
-    expect(ethereumAddress).not.toBeNull();
-
-    const originName = 'https://uniswap.org/'
-    if (typeof ethAccountId !== 'undefined' && typeof ethereumAddress !== 'undefined') {
-        // external connection
-        const externalConnection = {
-            origin: originName,
-            chain: ChainId.Ethereum,
-            accountId: ethAccountId,
-            setDefaultEthereum: true,
-        }
-        await wallet.dispatch.addExternalConnection(externalConnection);
-    }
-
-    expect(Object.keys(wallet.state.externalConnections[walletId]).length).toEqual(1);
-    expect(wallet.state.externalConnections[walletId]?.[originName]?.defaultEthereum).toEqual(ethAccountId)
-    expect(Object.keys(wallet.state.externalConnections[walletId]?.[originName]?.ethereum).length)
-        .toBeGreaterThan(0);
-    expect(wallet.state.externalConnections[walletId]?.[originName]?.ethereum[0]).toEqual(ethAccountId);
-
-    // forgot dapp connections
-    await wallet.dispatch.forgetDappConnections()
-    expect(Object.keys(wallet.state.externalConnections[walletId]).length).toEqual(0);
-})
